refactor(auction): event 패키지 auction 하위로 위치 변경
